+++ components/cards/CurrentPatientsCard.vue
@@ -1,0 +1,58 @@
+<template>
+  <v-col class="DataCard" cols="12" md="6">
+    <time-bar-chart
+      :chart-data="currentPatientsGraph"
+      :date="convertToDateFromData(current_patients.last_update)"
+      :default-data-kind="'cumulative'"
+      :loaded="current_patients.loaded"
+      :source-from="$t('北海道 オープンデータポータル')"
+      :supplement="
+        $t(
+          '現在患者数とは、陽性患者数から治療終了者数と死亡者数を除いた人数です。なお、ご覧いただいている時間によっては累計されている日付が違う場合がありますのでご注意ください。死亡者数は北海道のホームページを参照してください。'
+        )
+      "
+      :title="$t('現在患者数')"
+      :unit="$t('人')"
+      source-link="https://www.harp.lg.jp/opendata/dataset/1369.html"
+    />
+  </v-col>
+</template>
+
+<script>
+import TimeBarChart from '@/components/TimeBarChart.vue'
+import formatCurrentPatientsGraph from '@/utils/formatCurrentPatientsGraph'
+import convertToDateFromData from '@/utils/convertToDateFromData'
+export default {
+  name: 'CurrentPatientsCard',
+  components: {
+    TimeBarChart
+  },
+  data() {
+    return {
+      current_patients: {
+        loaded: false,
+        last_update: ''
+      },
+      currentPatientsGraph: [],
+      convertToDateFromData
+    }
+  },
+  created() {
+    this.getCurrentPatientsGraphFromAPI()
+  },
+  methods: {
+    async getCurrentPatientsGraphFromAPI() {
+      await this.$axios
+        .$get('/current_patients.json')
+        .then(response => {
+          this.currentPatientsGraph = formatCurrentPatientsGraph(response.data)
+          this.current_patients.last_update = response.last_update
+          this.current_patients.loaded = true
+        })
+        .catch(_ => {
+          this.$emit('failed', '現在患者数データ ')
+        })
+    }
+  }
+}
+</script>
