+++ pages/index.vue
@@ -12,106 +12,13 @@
     />
 
     <v-row class="DataBlock">
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('現在患者数')"
-          :chart-data="currentPatientsGraph"
-          :date="convertToDateFromData(current_patients.last_update)"
-          :source-from="$t('北海道 オープンデータポータル')"
-          source-link="https://www.harp.lg.jp/opendata/dataset/1369.html"
-          :loaded="current_patients.loaded"
-          :unit="$t('人')"
-          :default-data-kind="'cumulative'"
-          :supplement="
-            $t(
-              '現在患者数とは、陽性患者数から治療終了者数と死亡者数を除いた人数です。なお、ご覧いただいている時間によっては累計されている日付が違う場合がありますのでご注意ください。死亡者数は北海道のホームページを参照してください。'
-            )
-          "
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('治療終了者数')"
-          :chart-data="dischargesGraph"
-          :date="convertToDateFromData(discharges_summary.last_update)"
-          :source-from="$t('北海道 オープンデータポータル')"
-          source-link="https://www.harp.lg.jp/opendata/dataset/1369.html"
-          :unit="$t('人')"
-          :loaded="discharges_summary.loaded"
-          :default-data-kind="'cumulative'"
-          :supplement="
-            $t(
-              '治療終了者数とは道発表の「陰性確認済累計」と同じものです。「陰性確認済累計」とは、陽性の患者が軽快してから48時間後の1回目のPCR検査で陰性が確認され、それから12時間後の2回目のPCR検査でも陰性が確認された方の累計のことです。（3/9 鈴木知事のツイートから引用）'
-            )
-          "
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('陽性患者数')"
-          :chart-data="patientsGraph"
-          :date="convertToDateFromData(patients_summary.last_update)"
-          :source-from="$t('北海道 オープンデータポータル')"
-          source-link="https://www.harp.lg.jp/opendata/dataset/1369.html"
-          :unit="$t('人')"
-          :loaded="patients_summary.loaded"
-          :default-data-kind="'cumulative'"
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <data-table
-          :title="$t('陽性患者の属性')"
-          :chart-data="patientsTable"
-          :chart-option="{}"
-          :date="convertToDateFromData(patients.last_update)"
-          :source-from="$t('北海道 オープンデータポータル')"
-          :loaded="patients.loaded"
-          source-link="https://www.harp.lg.jp/opendata/dataset/1369.html"
-          :sort-by="'日付'"
-          :sort-desc="true"
-          :info="sumInfoOfPatients"
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('検査数')"
-          :chart-data="inspectionsGraph"
-          :date="convertToDateFromData(inspections.last_update)"
-          :source-from="$t('北海道 オープンデータポータル')"
-          source-link="https://www.harp.lg.jp/opendata/dataset/1369.html"
-          :unit="$t('人')"
-          :loaded="inspections.loaded"
-          :default-data-kind="'cumulative'"
-          :show-button="false"
-          :supplement="
-            $t(
-              '3月3日以前のデータが公開されていないため、グラフは3月3日以降となります。'
-            )
-          "
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('新型コロナコールセンター相談件数(札幌市保健所値)')"
-          :chart-data="contactsGraph"
-          :date="convertToDateFromData(contacts.last_update)"
-          :source-from="$t('DATA-SMART CITY SAPPORO')"
-          source-link="https://ckan.pf-sapporo.jp/dataset/covid_19_soudan"
-          :loaded="contacts.loaded"
-          :unit="$t('件')"
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('帰国者・接触者電話相談センター相談件数(札幌市保健所値)')"
-          :chart-data="querentsGraph"
-          :date="convertToDateFromData(querents.last_update)"
-          :source-from="$t('DATA-SMART CITY SAPPORO')"
-          source-link="https://ckan.pf-sapporo.jp/dataset/covid_19_soudan"
-          :loaded="querents.loaded"
-          :unit="$t('件')"
-        />
-      </v-col>
+      <current-patients-card />
+      <discharges-card />
+      <patients-summary-card />
+      <patients-card />
+      <inspections-card />
+      <contacts-card />
+      <querents-card />
     </v-row>
     <v-dialog v-model="failed" flat>
       <v-alert type="error">
@@ -126,266 +33,52 @@
 
 <script>
 import PageHeader from '@/components/PageHeader.vue'
-import TimeBarChart from '@/components/TimeBarChart.vue'
 import WhatsNew from '@/components/WhatsNew.vue'
-import DataTable from '@/components/DataTable.vue'
-import formatGraph from '@/utils/formatGraph'
-import formatTable from '@/utils/formatTable'
-import convertToDateFromData from '@/utils/convertToDateFromData'
-import formatCurrentPatientsGraph from '@/utils/formatCurrentPatientsGraph'
-import formatDischargesSummaryGraph from '@/utils/formatDischargesSummaryGraph'
-import formatInspectionsGraph from '@/utils/formatInspectionsGraph'
-import formatPatientsSummaryGraph from '@/utils/formatPatientsSummaryGraph'
-
-const axiosOptions = {}
+import convertToDateFromData from '@/utils/convertToDateFromData.ts'
+import DischargesCard from '@/components/cards/DischargesCard.vue'
+import PatientsSummaryCard from '@/components/cards/PatientsSummaryCard.vue'
+import PatientsCard from '@/components/cards/PatientsCard.vue'
+import ContactsCard from '@/components/cards/ContactsCard.vue'
+import QuerentsCard from '@/components/cards/QuerentsCard.vue'
+import CurrentPatientsCard from '@/components/cards/CurrentPatientsCard.vue'
+import InspectionsCard from '@/components/cards/InspectionsCard.vue'
 
 export default {
   components: {
+    InspectionsCard,
+    CurrentPatientsCard,
+    QuerentsCard,
+    ContactsCard,
+    PatientsCard,
+    PatientsSummaryCard,
+    DischargesCard,
     PageHeader,
-    TimeBarChart,
-    WhatsNew,
-    DataTable
+    WhatsNew
   },
   data() {
     const data = {
-      contacts: {
-        loaded: false,
-        last_update: ''
-      },
-      current_patients: {
-        loaded: false,
-        last_update: ''
-      },
-      discharges_summary: {
-        loaded: false,
-        last_update: ''
-      },
-      inspections: {
-        loaded: false,
-        last_update: ''
-      },
-      patients: {
-        loaded: false,
-        last_update: ''
-      },
-      patients_summary: {
-        loaded: false,
-        last_update: ''
-      },
-      querents: {
-        loaded: false,
-        last_update: ''
-      },
       /**
        * 全体の最終更新日
        */
       last_update: '',
-      /**
-       * 各グラフ系のデータ整理後のデータ
-       */
-      patientsTable: {},
-      patientsGraph: [],
-      contactsGraph: [],
-      querentsGraph: [],
-      currentPatientsGraph: [],
-      dischargesGraph: [],
-      inspectionsGraph: [],
-      failed: false,
-      failed_datas: '',
-      sumInfoOfPatients: {
-        lText: '',
-        sText: '',
-        unit: ''
-      },
       headerItem: {
         icon: 'mdi-chart-timeline-variant',
         title: this.$t('道内の最新感染動向'),
         date: ''
       },
-      option: {
-        responsive: true,
-        legend: {
-          display: false
-        },
-        scales: {
-          xAxes: [
-            {
-              stacked: true,
-              gridLines: {
-                display: false
-              },
-              ticks: {
-                fontSize: 10,
-                maxTicksLimit: 20,
-                fontColor: '#808080'
-              }
-            }
-          ],
-          yAxes: [
-            {
-              location: 'bottom',
-              stacked: true,
-              gridLines: {
-                display: true,
-                color: '#E5E5E5'
-              },
-              ticks: {
-                suggestedMin: 0,
-                maxTicksLimit: 8,
-                fontColor: '#808080'
-              }
-            }
-          ]
-        }
-      },
+      failed: false,
+      failed_datas: '',
       convertToDateFromData
     }
     return data
   },
   created() {
-    this.getContactsGraphFromAPI()
-    this.getCurrentPatientsGraphFromAPI()
-    this.getDischargesSummaryGraphFromAPI()
-    this.getInscpectionsGraphFromAPI()
     this.getLastUpdateFromAPI()
-    this.getPatientsTableFromAPI()
-    this.getPatientsSummaryGraphFromAPI()
-    this.getQuerentsGraphFromAPI()
   },
   methods: {
-    // 現在患者数グラフ
-    async getCurrentPatientsGraphFromAPI() {
-      await this.$axios
-        .$get('/current_patients.json', axiosOptions)
-        .then(response => {
-          this.currentPatientsGraph = formatCurrentPatientsGraph(response.data)
-          this.current_patients.last_update = response.last_update
-          this.current_patients.loaded = true
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '現在患者数データ '
-        })
-    },
-    // 感染者数グラフ
-    async getPatientsSummaryGraphFromAPI() {
-      await this.$axios
-        .$get('/patients_summary.json', axiosOptions)
-        .then(response => {
-          this.patientsGraph = formatPatientsSummaryGraph(response.data)
-          this.patients_summary.last_update = response.last_update
-          this.patients_summary.loaded = true
-          this.sumInfoOfPatients = {
-            lText: this.patientsGraph[
-              this.patientsGraph.length - 1
-            ].cumulative.toLocaleString(),
-            sText: this.$t('{date}の累計', {
-              date: this.$moment(
-                this.patientsGraph[this.patientsGraph.length - 1].label
-              ).format('MM/DD')
-            }),
-            unit: this.$t('人')
-          }
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '陽性患者数データ '
-        })
-    },
-    // 感染者
-    async getPatientsTableFromAPI() {
-      await this.$axios
-        .$get('/patients.json', axiosOptions)
-        .then(response => {
-          this.patientsTable = formatTable(response.data)
-          // 陽性患者の属性 ヘッダー翻訳
-          for (const header of this.patientsTable.headers) {
-            header.text =
-              header.value === '退院' ? this.$t('退院※') : this.$t(header.value)
-          }
-          // 陽性患者の属性 中身の翻訳
-          for (const row of this.patientsTable.datasets) {
-            row['居住地'] = this.$t(row['居住地'])
-            row['性別'] = this.$t(row['性別'])
-            row['退院'] = this.$t(row['退院'])
-            if (row['性別'] === '非公表性') {
-              row['性別'] = this.$t('非公表')
-            }
-            const match = row['年代'].match(/^([0-9]+)代$/)
-            if (match) {
-              row['年代'] = this.$t('{age}代', { age: match[1] })
-            } else {
-              row['年代'] = this.$t(row['年代'])
-            }
-          }
-          this.patients.last_update = response.last_update
-          this.patients.loaded = true
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '陽性患者の属性データ '
-        })
-    },
-    // 治療終了者数グラフ
-    async getDischargesSummaryGraphFromAPI() {
-      await this.$axios
-        .$get('/discharges_summary.json', axiosOptions)
-        .then(response => {
-          this.dischargesGraph = formatDischargesSummaryGraph(response.data)
-          this.discharges_summary.last_update = response.last_update
-          this.discharges_summary.loaded = true
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '治療終了者数データ '
-        })
-    },
-    // 検査数グラフ
-    async getInscpectionsGraphFromAPI() {
-      await this.$axios
-        .$get('/inspections.json', axiosOptions)
-        .then(response => {
-          this.inspectionsGraph = formatInspectionsGraph(response.data)
-          this.inspections.last_update = response.last_update
-          this.inspections.loaded = true
-          return true
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '検査数データ '
-        })
-    },
-    // 相談件数グラフ
-    async getContactsGraphFromAPI() {
-      await this.$axios
-        .$get('/contacts.json', axiosOptions)
-        .then(response => {
-          this.contactsGraph = formatGraph(response.data)
-          this.contacts.last_update = response.last_update
-          this.contacts.loaded = true
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '新型コロナコールセンター相談件数データ '
-        })
-    },
-    // 帰国者・接触者電話相談センター相談件数グラフ
-    async getQuerentsGraphFromAPI() {
-      await this.$axios
-        .$get('/querents.json', axiosOptions)
-        .then(response => {
-          this.querentsGraph = formatGraph(response.data)
-          this.querents.last_update = response.last_update
-          this.querents.loaded = true
-        })
-        .catch(_ => {
-          this.failed = true
-          this.failed_datas += '帰国者・接触者電話相談センター相談件数データ '
-        })
-    },
     async getLastUpdateFromAPI() {
       await this.$axios
-        .$get('/last_update.json', axiosOptions)
+        .$get('/last_update.json')
         .then(response => {
           this.headerItem = {
             icon: 'mdi-chart-timeline-variant',
